day-05-deploy-a-stateful-service-with-stateful-sets-and-persistent-volume-claims/app/config.py
from pydantic_settings import BaseSettings
from dotenv import load_dotenv
import os


load_dotenv()


user = os.getenv("DB_USERNAME")
password = os.getenv("DB_PASSWORD")


class Settings(BaseSettings):
    sqlalchemy_string: str = f"postgresql://{user}:{password}@host/chingae"
    
settings = Settings()